+++ CHANGELOG.md
@@ -15,6 +15,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Display handle with '@' in account link tooltips.
 - Change tooltip text on "Like" and "Repost" buttons after completing action.
+- Inline all SVG icons.
 
 ### Removed